requirements.txt
pygbif==0.4.0
pandas==0.25.3
PyYAML==5.3.1
